+++ tests/regressions.rs
@@ -1,13 +1,17 @@
 use std::io::{self, Read};
+use std::sync::Arc;
 use std::time::Duration;
 use std::{net, thread};
 
-use mio::net::TcpStream;
-use mio::{Events, Interest, Poll, Token};
+use mio::net::{TcpListener, TcpStream};
+use mio::{Events, Interest, Poll, Token, Waker};
 
 mod util;
 
-use util::init;
+use util::{any_local_address, init, init_with_poll};
+
+const ID1: Token = Token(1);
+const WAKE_TOKEN: Token = Token(10);
 
 #[test]
 fn issue_776() {
@@ -49,4 +53,43 @@ fn issue_776() {
 
     drop(s);
     t.join().unwrap();
+}
+
+#[test]
+fn issue_1205() {
+    let (mut poll, mut events) = init_with_poll();
+
+    let waker = Arc::new(Waker::new(poll.registry(), WAKE_TOKEN).unwrap());
+    let waker1 = waker.clone();
+
+    let mut listener = TcpListener::bind(any_local_address()).unwrap();
+
+    poll.registry()
+        .register(&mut listener, ID1, Interest::READABLE)
+        .unwrap();
+
+    poll.poll(&mut events, Some(std::time::Duration::from_millis(0)))
+        .unwrap();
+    assert!(events.iter().count() == 0);
+
+    let _stream = TcpStream::connect(listener.local_addr().unwrap()).unwrap();
+
+    poll.registry().deregister(&mut listener).unwrap();
+
+    // spawn a waker thread to wake the poll call below
+    let handle = thread::spawn(move || {
+        thread::sleep(Duration::from_millis(500));
+        waker1.wake().expect("unable to wake");
+    });
+
+    poll.poll(&mut events, None).unwrap();
+
+    // the poll should return only one event that being the waker event.
+    // the poll should not retrieve event for the listener above because it was
+    // deregistered
+    assert!(events.iter().count() == 1);
+    let waker_event = events.iter().next().unwrap();
+    assert!(waker_event.is_readable());
+    assert_eq!(waker_event.token(), WAKE_TOKEN);
+    handle.join().unwrap();
 }
